Add pacemaker integration code into VSM puppetry
This patch adds in the option to support pacemaker integration in OSP7.
Specifically when pacemaker integration is enabled two VMs will be
created- one for the primary, and one for the secondary. Neither VM
will be defined or started. Addditionally MACs will be generated for
both VMs to prevent an issue on VSM switchover.

Code was also added to support cases where the VSM needs to be
brought up in setups where there's an existing bridge.

Change-Id: I2ca2ccc12a9db56348440d210eef4bde25142aef